+++ ceammc/ext/src/ui/ui_gain.h
@@ -55,6 +55,7 @@ public:
     void onMouseDown(t_object* view, const t_pt& pt, const t_pt& abs_pt, long modifiers);
     void onMouseDrag(t_object* view, const t_pt& pt, long modifiers);
     void onDblClick(t_object* view, const t_pt& pt, long modifiers);
+    void onMouseWheel(const t_pt& pt, long modifiers, float delta);
 
     t_float dbValue() const;
     t_float ampValue() const;