+++ ceammc/CHANGELOG.md
@@ -22,11 +22,13 @@
 - ui.hgain2~ alias added to create horizontal ui.gain2~
 - ui.gain~ and ui.gain2~ @relative property added for relative input mode (by default)
 - ui.gain~ and ui.gain2~ slow value change added in relative mode while holding SHIFT
+- ui.gain~/ui.gain2~ mouse wheel support added (with slow on SHIFT)
 - ui.polar added
 - ui.slider slow change with SHIFT added
 - ui.slider log scale support added with @scale property
 - ui.knob slow value change with SHIFT added
 - ui.knob log scale support added with @scale property
+- ui.knob mouse wheel support added (with slow on SHIFT)
 - popup menu added to ui.preset with read/write actions
 - math sync binary operations added (all inlets are hot):
   - math.sync_add (math.+', +')